# Conflicts:
#	ibuypr/db.sqlite3

+++ ibuypr/ibuy/views.py
@@ -96,7 +96,7 @@ def criarconta(request):
                 FileSystemStorage().save('images/utilizador/' + nome_imagem, image)
                 utilizador.imagem = nome_imagem
             utilizador.save()
-            return HttpResponseRedirect(reverse('ibuy:index'))
+            return HttpResponseRedirect(reverse('ibuy:loginuser'))
     else:
         return render(request, 'ibuy/criarconta.html', context)
 
@@ -128,7 +128,7 @@ def logoutview(request):
 def minhaconta(request, user_id):
     user = get_object_or_404(User, pk=user_id)
     if not (request.user.is_superuser or user.id == request.user.id):
-        loginuser(request)
+        return render(request, 'ibuy/erro.html')
     user_form = UserForm(instance=user)
     context = {
         'user': user,
@@ -143,7 +143,7 @@ def minhaconta(request, user_id):
 def alterarpassword(request, user_id):
     user = get_object_or_404(User, pk=user_id)
     if not (request.user.is_superuser or user.id == request.user.id):
-        loginuser(request)
+        return render(request, 'ibuy/erro.html')
     user_form = UserForm(instance=user)
     if request.method == 'POST':
         passwordatual = request.POST['passwordatual']
@@ -173,7 +173,7 @@ def alterarpassword(request, user_id):
 def alterarconta(request, user_id):
     user = get_object_or_404(User, pk=user_id)
     if not (request.user.is_superuser or user.id == request.user.id):
-        loginuser(request)
+        return render(request, 'ibuy/erro.html')
     user_form = UserForm(instance=user)
     context = { #adicionar este onde necessário
         'user_form': user_form,
@@ -211,6 +211,8 @@ def alterarconta(request, user_id):
 
 def perfil(request, user_id):
     user = get_object_or_404(User, pk=user_id)
+    if user.is_superuser:
+        return HttpResponseRedirect(reverse('ibuy:erro'))
     lista_produtos = Produto.objects.filter(user_id=user.id)
     context = {
         'user': user,
@@ -366,7 +368,7 @@ def criarproduto(request):
         condicao = request.POST['condicao']
         categoria_id = request.POST['categoria']
         video_embed = request.POST['video_embed']
-        image = request.FILES.get('img_produto', False)
+        image = request.FILES.get('imagem', False)
         categoria = get_object_or_404(Categoria, pk=categoria_id)
         if not (nome and quantidade and preco and descricao and condicao and categoria):
             return render(request, 'ibuy/criarproduto.html',
@@ -394,7 +396,8 @@ def criarproduto(request):
 @login_required(login_url=reverse_lazy('ibuy:loginuser'))
 def apagarproduto(request, produto_id):
     produto = get_object_or_404(Produto, pk=produto_id)
-    FileSystemStorage().delete('images/produto/' + produto.imagem)
+    if not produto.imagem == 'produto.png':
+        FileSystemStorage().delete('images/produto/' + produto.imagem)
     produto.delete()
     return HttpResponseRedirect(reverse('ibuy:meusprodutos'))
 
@@ -508,7 +511,7 @@ def alterarproduto(request, produto_id):
         condicao = request.POST['condicao']
         categoria_id = request.POST['categoria']
         categoria = get_object_or_404(Categoria, pk=categoria_id)
-        image = request.FILES.get('img_produto', False)
+        image = request.FILES.get('imagem', False)
         video_embed = request.POST['video_embed']
 
         if not (nome and quantidade and preco and descricao and condicao and categoria):
@@ -559,7 +562,8 @@ def utilizadores(request):
 @user_passes_test(is_admin, login_url=reverse_lazy('ibuy:erro'))
 def apagarutilizador(request, user_id):
     user = get_object_or_404(User, pk=user_id)
-    FileSystemStorage().delete('images/utilizador/' + user.utilizador.imagem)
+    if not user.utilizador.imagem == 'utilizador.png':
+        FileSystemStorage().delete('images/utilizador/' + user.utilizador.imagem)
     user.utilizador.delete()
     user.delete()
     return HttpResponseRedirect(reverse('ibuy:utilizadores'))
@@ -580,6 +584,11 @@ def historiaempresa(request):
 
 def ondeestamos(request):
     return render(request, 'ibuy/ondeestamos.html')
+
+
+def nossamoeda(request):
+    return render(request, 'ibuy/nossamoeda.html')
+
 
 def adicionarmoeda(request):
     return render(request, 'ibuy/adicionarmoeda.html')
